+++ markup/Readme.md
@@ -1,10 +1,10 @@
 # Директория с вёрсткой (markup)
 
 * `index.html` - контейнер;
-* `login.html` - страница «Login» (Авторизация);
+* `login.html` - страница «Index» (Авторизация);
 * `main.html` - страница «Main» (Главная страница со списком предложений):
    * `main-empty.html` - без списка предложений;
 * `property.html` — страница «Room» (Карточка предложения по аренде, пользователь авторизован):
    * `property-not-logged.html` - пользователь не авторизован;
-* `favorites.html` - страница «Favorites» (Избранное):
+* `favorites.html` - страница «Index» (Избранное):
    * `favorites-empty.html` - пустая страница.
