+++ markup/favorites-empty.html
@@ -46,7 +46,7 @@
       <main class="page__main page__main--favorites page__main--favorites-empty">
         <div class="page__favorites-container container">
           <section class="favorites favorites--empty">
-            <h1 class="visually-hidden">Favorites (empty)</h1>
+            <h1 class="visually-hidden">Index (empty)</h1>
             <div class="favorites__status-wrapper">
               <b class="favorites__status">Nothing yet saved.</b>
               <p class="favorites__status-description">Save properties to narrow down search or plan your future trips.</p>
