+++ project/src/mocks/offers.ts
@@ -1,37 +1,37 @@
 export const offers = [
   {
-    title: `Beautiful & luxurious apartment at great location`,
-    id: 1,
+    title: 'Beautiful & luxurious apartment at great location',
+    id: '1',
     price: 120,
-    type: `Apartment`,
-    imgSrc: `img/apartment-01.jpg`,
+    type: 'Apartment',
+    imgSrc: 'img/apartment-01.jpg',
     rating: 4,
     isPremium: true,
   },
   {
-    title: `Wood and stone place`,
-    id: 2,
+    title: 'Wood and stone place',
+    id: '2',
     price: 132,
-    type: `Private room`,
-    imgSrc: `img/room.jpg`,
+    type: 'Private room',
+    imgSrc: 'img/room.jpg',
     rating: 4,
     isPremium: false,
   },
   {
-    title: `Canal View Prinsengracht`,
-    id: 3,
-    price: 120,
-    type: `Apartment`,
-    imgSrc: `img/apartment-02.jpg`,
+    title: 'Canal View Prinsengracht',
+    id: '3',
+    price: 140,
+    type: 'Apartment',
+    imgSrc: 'img/apartment-02.jpg',
     rating: 5,
     isPremium: true,
   },
   {
-    title: `Nice, cozy, warm big bed apartment`,
-    id: 4,
+    title: 'Nice, cozy, warm big bed apartment',
+    id: '4',
     price: 80,
-    type: `apartment`,
-    imgSrc: `img/apartment-03.jpg`,
+    type: 'apartment',
+    imgSrc: 'img/apartment-03.jpg',
     rating: 3,
     isPremium: false,
   },
