+++ project/src/pages/main/index.tsx
@@ -1,10 +1,13 @@
-import {Card} from '../../components/card/Card';
+import {CardProps} from '../../components/App/App';
+import {OffersList} from '../../components/OffersList/OffersList';
+import {FC} from 'react';
 
 type MainScreenProps = {
-  cardsCount: number
+  cardsCount: number,
+  offers: CardProps[]
 }
 
-export const MainScreen = ({cardsCount}: MainScreenProps): JSX.Element => (
+export const MainScreen:FC<MainScreenProps> = ({cardsCount, offers}) => (
   <div className="page page--gray page--main">
     <header className="header">
       <div className="container">
@@ -93,13 +96,7 @@ export const MainScreen = ({cardsCount}: MainScreenProps): JSX.Element => (
                 <li className="places__option" tabIndex={0}>Top rated first</li>
               </ul>
             </form>
-            <div className="cities__places-list places__list tabs__content">
-              <Card/>
-              <Card/>
-              <Card/>
-              <Card/>
-              <Card/>
-            </div>
+            <OffersList content={offers} cardsCount={cardsCount}/>
           </section>
           <div className="cities__right-section">
             <section className="cities__map map"/>
