+++ project/src/pages/offer/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export const PropertyNotLogged = (): JSX.Element => (
+export const Index = (): JSX.Element => (
   <div className="page">
     <header className="header">
       <div className="container">
@@ -16,7 +16,13 @@ export const PropertyNotLogged = (): JSX.Element => (
                 <a className="header__nav-link header__nav-link--profile" href="#">
                   <div className="header__avatar-wrapper user__avatar-wrapper">
                   </div>
-                  <span className="header__login">Sign in</span>
+                  <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
+                  <span className="header__favorite-count">3</span>
+                </a>
+              </li>
+              <li className="header__nav-item">
+                <a className="header__nav-link" href="#">
+                  <span className="header__signout">Sign out</span>
                 </a>
               </li>
             </ul>
@@ -60,14 +66,14 @@ export const PropertyNotLogged = (): JSX.Element => (
               </h1>
               <button className="property__bookmark-button button" type="button">
                 <svg className="property__bookmark-icon" width="31" height="33">
-                  <use xlinkHref="#icon-bookmark" />
+                  <use xlinkHref="#icon-bookmark"/>
                 </svg>
                 <span className="visually-hidden">To bookmarks</span>
               </button>
             </div>
             <div className="property__rating rating">
               <div className="property__stars rating__stars">
-                <span style={{width: '80%'}} />
+                <span style={{width: '80%'}}/>
                 <span className="visually-hidden">Rating</span>
               </div>
               <span className="property__rating-value rating__value">4.8</span>
@@ -165,7 +171,7 @@ export const PropertyNotLogged = (): JSX.Element => (
                   <div className="reviews__info">
                     <div className="reviews__rating rating">
                       <div className="reviews__stars rating__stars">
-                        <span style={{width: '80%'}} />
+                        <span style={{width: '80%'}}/>
                         <span className="visually-hidden">Rating</span>
                       </div>
                     </div>
@@ -177,10 +183,71 @@ export const PropertyNotLogged = (): JSX.Element => (
                   </div>
                 </li>
               </ul>
+              <form className="reviews__form form" action="#" method="post">
+                <label className="reviews__label form__label" htmlFor="review">Your review</label>
+                <div className="reviews__rating-form form__rating">
+                  <input className="form__rating-input visually-hidden" name="rating" value="5" id="5-stars"
+                    type="radio"
+                  />
+                  <label htmlFor="5-stars" className="reviews__rating-label form__rating-label" title="perfect">
+                    <svg className="form__star-image" width="37" height="33">
+                      <use xlinkHref="#icon-star"/>
+                    </svg>
+                  </label>
+
+                  <input className="form__rating-input visually-hidden" name="rating" value="4" id="4-stars"
+                    type="radio"
+                  />
+                  <label htmlFor="4-stars" className="reviews__rating-label form__rating-label" title="good">
+                    <svg className="form__star-image" width="37" height="33">
+                      <use xlinkHref="#icon-star"/>
+                    </svg>
+                  </label>
+
+                  <input className="form__rating-input visually-hidden" name="rating" value="3" id="3-stars"
+                    type="radio"
+                  />
+                  <label htmlFor="3-stars" className="reviews__rating-label form__rating-label" title="not bad">
+                    <svg className="form__star-image" width="37" height="33">
+                      <use xlinkHref="#icon-star"/>
+                    </svg>
+                  </label>
+
+                  <input className="form__rating-input visually-hidden" name="rating" value="2" id="2-stars"
+                    type="radio"
+                  />
+                  <label htmlFor="2-stars" className="reviews__rating-label form__rating-label" title="badly">
+                    <svg className="form__star-image" width="37" height="33">
+                      <use xlinkHref="#icon-star"/>
+                    </svg>
+                  </label>
+
+                  <input className="form__rating-input visually-hidden" name="rating" value="1" id="1-star"
+                    type="radio"
+                  />
+                  <label htmlFor="1-star" className="reviews__rating-label form__rating-label"
+                    title="terribly"
+                  >
+                    <svg className="form__star-image" width="37" height="33">
+                      <use xlinkHref="#icon-star"/>
+                    </svg>
+                  </label>
+                </div>
+                <textarea className="reviews__textarea form__textarea" id="review" name="review"
+                  placeholder="Tell how was your stay, what you like and what can be improved"
+                />
+                <div className="reviews__button-wrapper">
+                  <p className="reviews__help">
+                      To submit review please make sure to set <span className="reviews__star">rating</span> and
+                      describe your stay with at least <b className="reviews__text-amount">50 characters</b>.
+                  </p>
+                  <button className="reviews__submit form__submit button" type="submit" disabled>Submit</button>
+                </div>
+              </form>
             </section>
           </div>
         </div>
-        <section className="property__map map" />
+        <section className="property__map map"/>
       </section>
       <div className="container">
         <section className="near-places places">
@@ -189,7 +256,7 @@ export const PropertyNotLogged = (): JSX.Element => (
             <article className="near-places__card place-card">
               <div className="near-places__image-wrapper place-card__image-wrapper">
                 <a href="#">
-                  <img className="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place image" />
+                  <img className="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place image"/>
                 </a>
               </div>
               <div className="place-card__info">
@@ -202,14 +269,14 @@ export const PropertyNotLogged = (): JSX.Element => (
                     type="button"
                   >
                     <svg className="place-card__bookmark-icon" width="18" height="19">
-                      <use xlinkHref="#icon-bookmark" />
+                      <use xlinkHref="#icon-bookmark"/>
                     </svg>
                     <span className="visually-hidden">In bookmarks</span>
                   </button>
                 </div>
                 <div className="place-card__rating rating">
                   <div className="place-card__stars rating__stars">
-                    <span style={{width: '80%'}} />
+                    <span style={{width: '80%'}}/>
                     <span className="visually-hidden">Rating</span>
                   </div>
                 </div>
@@ -236,14 +303,14 @@ export const PropertyNotLogged = (): JSX.Element => (
                   </div>
                   <button className="place-card__bookmark-button button" type="button">
                     <svg className="place-card__bookmark-icon" width="18" height="19">
-                      <use xlinkHref="#icon-bookmark" />
+                      <use xlinkHref="#icon-bookmark"/>
                     </svg>
                     <span className="visually-hidden">To bookmarks</span>
                   </button>
                 </div>
                 <div className="place-card__rating rating">
                   <div className="place-card__stars rating__stars">
-                    <span style={{width: '80%'}} />
+                    <span style={{width: '80%'}}/>
                     <span className="visually-hidden">Rating</span>
                   </div>
                 </div>
@@ -260,9 +327,7 @@ export const PropertyNotLogged = (): JSX.Element => (
               </div>
               <div className="near-places__image-wrapper place-card__image-wrapper">
                 <a href="#">
-                  <img className="place-card__image" src="img/apartment-03.jpg" width="260" height="200"
-                    alt="Place image"
-                  />
+                  <img className="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place image"/>
                 </a>
               </div>
               <div className="place-card__info">
@@ -273,14 +338,14 @@ export const PropertyNotLogged = (): JSX.Element => (
                   </div>
                   <button className="place-card__bookmark-button button" type="button">
                     <svg className="place-card__bookmark-icon" width="18" height="19">
-                      <use xlinkHref="#icon-bookmark" />
+                      <use xlinkHref="#icon-bookmark"/>
                     </svg>
                     <span className="visually-hidden">To bookmarks</span>
                   </button>
                 </div>
                 <div className="place-card__rating rating">
                   <div className="place-card__stars rating__stars">
-                    <span style={{width: '100%'}} />
+                    <span style={{width: '100%'}}/>
                     <span className="visually-hidden">Rating</span>
                   </div>
                 </div>
